int32_t

+++ include/base/container/IQueue.h
@@ -12,7 +12,7 @@ namespace base
 		virtual ~IQueue() = default;
 
 	public:
-		virtual uint64_t Count() const = 0;
+		virtual int32_t Count() const = 0;
 		virtual T Dequeue() = 0;
 		virtual bool TryDequeue(T &out) = 0;
 		virtual void Enqueue(T obj) = 0;
